Improve and debug the case that missing state diagram
