+++ src/Model/oop/ProcessList.java
@@ -1,10 +1,10 @@
-package Model;
+package Model.oop;
 
 import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.Map;
 
-import Model.Proc.ProcessLinkedList;
+import Model.proc.ProcessLinkedList;
 
 public class ProcessList {
