+++ src/Model/state/StateProcess.java
@@ -1,4 +1,4 @@
-package Model;
+package Model.state;
 
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -7,7 +7,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 
-import Model.Proc.ProcessLinkedList;
+import Model.oop.ProcessList;
+import Model.proc.ProcessLinkedList;
 
 public class StateProcess extends ProcessList {
